TST: move np.testing.dec tests to test_deprecations.py
